[http-assert] change intended parameter types in favor of their real supported underlying types.
The functions of this module all expose the same series of parameters => assert(..., status, msg, opts):void

Those three parameters are intended to be use in a specific order, as their name suggest.

But http-assert passes those parameters directly to http-errors, who can handle the parameters in any order.

This commit widen the type signature to get it closer to the actual signature of CreateHttpError, without changing the current names and description of the indented use.